Make Ogg music track names configurable
With "OggTrackName" in sw.cfg, so you can for example
set it to "MUSIC/Track??.ogg" for the SW Steam version
"??" is used as a placeholder for the actual track number,
apart from that any path up to 255 chars (relative to the
base SW directory) can be configured.
It defaults to "track??.ogg", like before.