Make (Quick)Save menu default to last loaded game
When you load a game from the menu, the Quicksave menu (F6)
selected the last saved game (or, when none has been saved yet
in this session, the first game), instead of the game you loaded.
This bug/inconvenience was already there in the original DOS client.

Now it defaults to the last loaded game (if any).