make music toggling work during the initial demo cycle
Gives the SongLevelNum variable a purpose. It should stay a 0 while the
theme is played regardless of the demos loop's map changes, but in a
game it ought to follow the level being played. It could probably also
safely replace the LevelSong var to index into LevelInfo if I can't
find a reason why the filename needs to be copied to it or saved to the
savegame file.